Reduce and simplify build configuration
Removed the whole `pluginManagement` block in the settings file as the
configuration was not needed for this project.

Moved jgtiver to the `plugins` block in the build script and removed the
`getGitVersion` function. The version currently differs from the one
before. We can fix this later on.

Simplified configurations by removing the lwjgl ones and instead using
the splash source set configurations directly.

Simplified the api elements shenanigans by just using the `api` function
in the `dependencies` block.

Added some comments here and there to document why we need certain
stuff.

Signed-off-by: Niklas Wimmer <mail@nwimmer.me>